feat(changelog-summarizer): pre-render Block Kit and publish via webhook
- formatters/blockkit.ts (shared): buildBlocks() splits long sections under
  Slack's 3000-char limit, hard-splitting even a single long paragraph.
- index.ts emits the Block Kit payload (<output>.blocks.json) and folds the
  technical notes into a final "For engineers" section (webhook publishing has
  no threaded reply); exposes blocks_file via GITHUB_OUTPUT.
- Workflow posts the pre-rendered payload to an incoming webhook
  (SLACK_WEBHOOK_F0_SUMMARIZER) — drops the fragile bash block-building and the
  bot-token/chat.postMessage path whose secrets were never configured. Point
  the webhook at #f0-design-system from the "Zero Updater" (zerito) app.

Verified a quarter-to-date payload (238 PRs) stays within limits: 19 blocks,
max section 2888 chars.
